feat: implement standard newsletter signup component and show when unverified Ensures sign up is visible even if user chooses not to complete identity verification
