+++ package.json
@@ -17,7 +17,7 @@
     "@polkadot/util": "12.6.2",
     "@polkadot/util-crypto": "12.6.2",
     "@polymeshassociation/browser-extension-signing-manager": "^2.3.0",
-    "@polymeshassociation/polymesh-sdk": "28.0.0-beta.3",
+    "@polymeshassociation/polymesh-sdk": "^28.0.0",
     "@polymeshassociation/walletconnect-signing-manager": "^1.0.0",
     "@tanstack/react-table": "^8.20.5",
     "@types/lodash": "^4.17.15",
