+++ src/components/UserAuth/components/NewsletterSignup/helpers.ts
