+++ src/components/UserAuth/components/NewsletterSignup/styles.ts
@@ -1,5 +1,23 @@
 import styled from 'styled-components';
 
+export const StyledNewsContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  min-width: 648px;
+  & > .subscribe-btn {
+    width: 'fit-content';
+    margin-left: auto;
+  }
+`;
+
+export const StyledInputsContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+  margin-top: 8px;
+`;
+
 export const StyledEmailField = styled.div`
   display: flex;
   align-items: center;
@@ -33,4 +51,7 @@ export const StyledCheckBox = styled.div<{ $checked: boolean }>`
   border: 1px solid;
   border-color: ${({ theme, $checked }) =>
     $checked ? theme.colors.textPink : '#8f8f8f'};
+  & svg {
+    color: ${({ theme }) => theme.colors.landingBackground};
+  }
 `;
