--- src/components/UserAuth/components/ViewUnverified/styles.ts
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledAuthButtons = styled.div`
-  display: flex;
-  gap: 16px;
-  @media screen and (max-width: 520px) {
-    flex-direction: column;
-  }
-`;
